Code cleanup & API response caching
We kick off version 4.1.0 with a re-evaluation of our 4.0.5 code, re-writing it line by line.

This commit doesn't encompass all of the functionalities in 4.0.5, but does clean up the way we make Boards API calls. I've gone ahead and utilized API response caching to hopefully slightly improve rendering times and decrease load on the Boards API.